+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/toolbar/DefaultToolbarMenu.kt
@@ -12,6 +12,7 @@ import androidx.core.content.ContextCompat.getColor
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.flow.distinctUntilChangedBy
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.launch
 import mozilla.components.browser.menu.BrowserMenuHighlight
@@ -155,6 +156,8 @@ open class DefaultToolbarMenu(
         )
 
         registerForIsBookmarkedUpdates()
+
+        registerForScreenReaderUpdates()
 
         BrowserMenuItemToolbar(listOf(back, forward, share, refresh), isSticky = true)
     }
@@ -480,6 +483,25 @@ open class DefaultToolbarMenu(
             isCurrentUrlBookmarked = bookmarksStorage
                 .getBookmarksWithUrl(newUrl)
                 .any { it.url == newUrl }
+        }
+    }
+
+    private fun registerForScreenReaderUpdates() {
+        store.flowScoped(lifecycleOwner) { flow ->
+            flow.mapNotNull { state -> state.selectedTab }
+                .distinctUntilChangedBy { it.readerState }
+                .collect {
+                    translationsItem.enabled = !it.readerState.active
+                    translationsItem.iconTintColorResource =
+                        if (it.readerState.active) {
+                            ThemeManager.resolveAttribute(
+                                R.attr.textDisabled,
+                                context,
+                            )
+                        } else {
+                            primaryTextColor()
+                        }
+                }
         }
     }
 }